03_economic_outlook/requirements.txt
# フェーズ3: 経済指標見通し考察用ライブラリ
# インストール方法: pip install -r requirements.txt

# フェーズ1・2の基本ライブラリ
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# Webスクレイピング用
beautifulsoup4>=4.12.0
lxml>=4.9.0

# データ取得用
pandas-datareader>=0.10.0
yfinance>=0.2.0
fredapi>=0.5.0 